feat: add OrderStatus enum
We assign a value to each variable to prevent an inexperienced developer from adding a variable in the middle of the enumeration. Doing so would reorder the indices of subsequent variables. Manually assigning a value to each variable prevents this. When you explicitly declare an enum value, you also need to implement a constructor, as shown below.